% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmsc}
\alias{cmsc}
\alias{gcsm}
\title{Composite similarity between vectors}
\usage{
cmsc(
  x,
  y,
  rescale = FALSE,
  xmin = NA_real_,
  xmax = NA_real_,
  ymin = NA_real_,
  ymax = NA_real_,
  comp = "si"
)

gcsm(
  x,
  y,
  rescale = FALSE,
  xmin = NA_real_,
  xmax = NA_real_,
  ymin = NA_real_,
  ymax = NA_real_,
  comp = "si"
)
}
\arguments{
\item{x}{A vector.}

\item{y}{The other vector.}

\item{rescale}{Rescale or not before computation.}

\item{xmin, xmax, ymin, ymax}{Rescale parameters.
If \code{NA_real_}, will set as global minimum and maximum in \code{x} and \code{y}, respectively.}

\item{comp}{Variable to return.
If \code{"si"}, the composite measure, if \code{"s1"},\code{"s2"} or \code{"s3"}, the corresponding component.}
}
\value{
A number.
}
\description{
Compute composite measures, GCSM or CMSC, between two vectors.
}
\examples{
# mean shift
x = runif(9)
y = 1 - x # y is the perfect antianalog of x
## similarity
gcsm(x, x)
cmsc(x, x)
gcsm(x, x - 0.2, xmin = 0, xmax = 1, ymin = 0, ymax = 1)
cmsc(x, x - 0.2, xmin = 0, xmax = 1, ymin = 0, ymax = 1)
gcsm(x, x + 0.2, xmin = 0, xmax = 1, ymin = 0, ymax = 1)
cmsc(x, x + 0.2, xmin = 0, xmax = 1, ymin = 0, ymax = 1)
## dissimilarity
gcsm(y, x)
gcsm(y, x - 0.2, xmin = 0, xmax = 1, ymin = 0, ymax = 1)
gcsm(y, x + 0.2, xmin = 0, xmax = 1, ymin = 0, ymax = 1)

# random noise
noise = rnorm(9, mean = 0, sd = 0.2)
## similarity
gcsm(x, x + noise, xmin = 0, xmax = 1, ymin = 0, ymax = 1)
cmsc(x, x + noise, xmin = 0, xmax = 1, ymin = 0, ymax = 1)
## dissimilarity
gcsm(y, x + noise, xmin = 0, xmax = 1, ymin = 0, ymax = 1)

}
