% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmsc_tw}
\alias{cmsc_tw}
\alias{cmsc_e1_tw}
\alias{cmsc_e2_tw}
\alias{gcsm_tw}
\title{Composite similarity on temporal windows}
\usage{
cmsc_tw(
  xxx,
  yyy,
  rescale = FALSE,
  xmin = NA_real_,
  xmax = NA_real_,
  ymin = NA_real_,
  ymax = NA_real_,
  comp = "si"
)

cmsc_e1_tw(
  xxx,
  yyy,
  rescale = FALSE,
  xmin = NA_real_,
  xmax = NA_real_,
  ymin = NA_real_,
  ymax = NA_real_,
  comp = "si"
)

cmsc_e2_tw(
  xxx,
  yyy,
  rescale = FALSE,
  xmin = NA_real_,
  xmax = NA_real_,
  ymin = NA_real_,
  ymax = NA_real_,
  comp = "si"
)

gcsm_tw(
  xxx,
  yyy,
  rescale = FALSE,
  xmin = NA_real_,
  xmax = NA_real_,
  ymin = NA_real_,
  ymax = NA_real_,
  comp = "si"
)
}
\arguments{
\item{xxx}{A 3-d array with the 3rd dimension representing time.}

\item{yyy}{The other 3-d array.}

\item{rescale}{Rescale or not before computation.}

\item{xmin, xmax, ymin, ymax}{Rescale parameters.
If \code{NA_real_}, will set as global minimum and maximum in \code{xxx} and \code{yyy}, respectively.}

\item{comp}{Variable to return.
If \code{"si"}, the composite measure, if \code{"s1"},\code{"s2"} or \code{"s3"}, the corresponding component.}
}
\value{
A matrix.
}
\description{
Compute composite measures, GCSM or CMSC, on temporal windows.
}
\examples{
x = array(runif(81), dim = c(3, 3, 9))

gcsm_tw(x, x + 0.2, xmin = 0, xmax = 1, ymin = 0, ymax = 1)
cmsc_tw(x, x + 0.2, xmin = 0, xmax = 1, ymin = 0, ymax = 1)

}
